feat!: extract the editorial layer into zingo-viewmodel and funnel zingo-cli through it
The value-transfer derivation, the finsight rollups, and their types
move out of zingolib into the new zingo-viewmodel crate, reached
through extension traits that preserve every method name and output
shape. The crate is also the funnel of ADR 0024 rule 7 (amendment
2026-08-03): its modules re-export zingolib's consumer surface
path-for-path, and zingo-cli becomes the first tenant by renaming its
single wallet dependency. Three summary primitives are promoted to
pub for the cross-crate derivation; the relocated golden fixtures all
reproduce today's output byte-identically.
