+++ Cargo.toml
@@ -7,6 +7,7 @@ members = [
     "zingo-net-diag",
     "zingo-price",
     "zingo-status",
+    "zingo-viewmodel",
     "zingolib",
     "zingolib_testutils",
 ]
@@ -33,6 +34,7 @@ zingo-net-diag = { path = "zingo-net-diag" }
 zingo-price = { path = "zingo-price", default-features = false }
 zingo-status = { path = "zingo-status" } # NOTE: for development between releases
 # zingo-status = "0.2.1"
+zingo-viewmodel = { path = "zingo-viewmodel" }
 zingolib = { path = "zingolib" }
 
 ### Tests and Integration Tests
