+++ zingo-cli/src/main.rs
@@ -82,7 +82,7 @@ fn handle_error(e: std::io::Error) {
 }
 
 pub fn main() {
-    zingolib::netutils::ensure_default_crypto_provider();
+    zingo_viewmodel::netutils::ensure_default_crypto_provider();
     let matches = parse_args_or_exit_for_help();
     init_tracing(&matches);
     if let Err(e) = zingo_cli::run_cli(matches) {
