+++ zingo-viewmodel/Cargo.toml
@@ -1,0 +1,37 @@
+[package]
+name = "zingo-viewmodel"
+description = "Editorial reductions over zingolib's canonical wallet data, and the single dependency consumers funnel through."
+version = "0.1.0"
+authors = ["<zingo@zingolabs.org>"]
+edition = "2024"
+repository = "https://github.com/zingolabs/zingolib"
+homepage = "https://github.com/zingolabs/zingolib"
+license = "MIT"
+
+[features]
+# Forwarded feature gates: the funnel mirrors zingolib's surface, so each
+# zingolib feature a consumer could previously enable is re-exported here
+# under the same name (ADR 0024 rule 7, amendment 2026-08-03).
+nym = ["zingolib/nym"]
+nym-diary = ["zingolib/nym-diary"]
+testutils = ["zingolib/testutils", "dep:tracing"]
+
+[dependencies]
+chrono = { workspace = true }
+json = { workspace = true }
+tracing = { workspace = true, optional = true }
+zcash_protocol = { workspace = true }
+zingo-status = { workspace = true }
+zingolib = { workspace = true }
+
+[dev-dependencies]
+incrementalmerkletree = { workspace = true }
+orchard = { workspace = true }
+pepper-sync = { workspace = true, features = ["test-features"] }
+tokio = { workspace = true, features = ["full"] }
+zcash_client_backend = { workspace = true }
+zcash_primitives = { workspace = true }
+zingo_common_components = { workspace = true }
+zingo_test_vectors = { workspace = true }
+zingolib = { workspace = true, features = ["testutils"] }
+zip32 = { workspace = true }
