+++ zingolib/src/wallet/summary.rs
@@ -1,10 +1,8 @@
 //! Types and impls for conveniently displaying information to the user or converting to JSON for interfacing with a larger stack.
-use std::collections::HashSet;
 /// A "snapshot" of the state of the items in the wallet at the time the summary was constructed.
 /// Not to be used for internal logic in the system.
-use std::{cmp::Ordering, collections::HashMap};
+use std::cmp::Ordering;
 
-use zcash_protocol::PoolType;
 use zcash_protocol::memo::Memo;
 
 use pepper_sync::keys::transparent;
@@ -13,49 +11,14 @@ use pepper_sync::wallet::{
 };
 
 use super::LightWallet;
-use super::error::{KeyError, SummaryError};
-use data::finsight::{
-    TotalMemoBytesToAddress, TotalSendsToAddress, TotalValueToAddress, ValuesSentToAddress,
-};
+use super::error::SummaryError;
 use data::{
     BasicCoinSummary, BasicNoteSummary, CoinSummary, NoteSummaries, NoteSummary,
-    OutgoingCoinSummary, OutgoingNoteSummary, Scope, SelfSendValueTransfer, SendType,
-    SentValueTransfer, TransactionKind, TransactionSummaries, TransactionSummary, ValueTransfer,
-    ValueTransferKind, ValueTransfers, pools_present,
+    OutgoingCoinSummary, OutgoingNoteSummary, Scope, SendType, TransactionKind,
+    TransactionSummaries, TransactionSummary,
 };
 
 pub mod data;
-
-/// Creates one value transfer of `kind` for each shielded pool the transaction
-/// received notes into, newest pool first (ironwood, orchard, sapling).
-fn shielded_pool_value_transfers(
-    transaction: &TransactionSummary,
-    kind: ValueTransferKind,
-) -> Vec<ValueTransfer> {
-    transaction
-        .shielded_notes_by_pool()
-        .into_iter()
-        .filter(|(notes, _)| !notes.is_empty())
-        .map(|(notes, pool)| {
-            ValueTransfer::from_summary(
-                transaction,
-                kind,
-                notes.iter().map(|output| output.value).sum(),
-                None,
-                vec![pool],
-                notes.iter().filter_map(|note| note.memo.clone()).collect(),
-            )
-        })
-        .collect()
-}
-
-/// Pairs each note with `pool`, for chaining outgoing notes across pools.
-fn tag_pool(
-    notes: &[OutgoingNoteSummary],
-    pool: PoolType,
-) -> impl Iterator<Item = (&OutgoingNoteSummary, PoolType)> {
-    notes.iter().map(move |note| (note, pool))
-}
 
 impl LightWallet {
     /// Returns summaries of all transactions in the wallet, sorted by confirmation status (confirmed first) and then
@@ -324,101 +287,6 @@ impl LightWallet {
         Ok(TransactionSummaries::new(transaction_summaries))
     }
 
-    /// Provides a list of value transfers related to this capability.
-    /// A value transfer is a group of all notes to a specific receiver in a transaction.
-    pub async fn value_transfers(
-        &self,
-        sort_highest_to_lowest: bool,
-    ) -> Result<ValueTransfers, SummaryError> {
-        let mut value_transfers: Vec<ValueTransfer> = Vec::new();
-        let transaction_summaries = self.transaction_summaries(sort_highest_to_lowest).await?.0;
-
-        for transaction in transaction_summaries {
-            match transaction.kind {
-                TransactionKind::Sent(SendType::Send) => {
-                    // create 1 sent value transfer for each non-self recipient address
-                    // if recipient_ua is available it overrides recipient_address
-                    value_transfers.append(&mut self.create_send_value_transfers(&transaction)?);
-
-                    // create 1 memo-to-self if any number of memos are received in the sending transaction
-                    let memos = transaction.received_memos();
-                    if !memos.is_empty() {
-                        value_transfers.push(ValueTransfer::from_summary(
-                            &transaction,
-                            ValueTransferKind::Sent(SentValueTransfer::SendToSelf(
-                                SelfSendValueTransfer::MemoToSelf,
-                            )),
-                            0,
-                            None,
-                            transaction.pools_received(),
-                            memos,
-                        ));
-                    }
-                }
-                TransactionKind::Sent(SendType::Shield) => {
-                    // create 1 shielding value transfer for each pool shielded to
-                    value_transfers.extend(shielded_pool_value_transfers(
-                        &transaction,
-                        ValueTransferKind::Sent(SentValueTransfer::SendToSelf(
-                            SelfSendValueTransfer::Shield,
-                        )),
-                    ));
-                }
-                TransactionKind::Sent(SendType::SendToSelf) => {
-                    // create 1 migration transfer if this moved Orchard funds into the Ironwood
-                    // pool (migration parts self-receive an empty Ironwood memo, so this must win
-                    // over the memo check), else 1 memo-to-self if a sending transaction receives
-                    // any number of memos, otherwise 1 basic send-to-self so every transaction
-                    // creates at least 1 value transfer.
-                    // (deshield and other pool-movement kinds may join this list later.)
-                    let memos = transaction.received_memos();
-                    let self_send_kind = if transaction.is_orchard_to_ironwood_migration() {
-                        SelfSendValueTransfer::Migration
-                    } else if !memos.is_empty() {
-                        SelfSendValueTransfer::MemoToSelf
-                    } else {
-                        SelfSendValueTransfer::Basic
-                    };
-                    value_transfers.push(ValueTransfer::from_summary(
-                        &transaction,
-                        ValueTransferKind::Sent(SentValueTransfer::SendToSelf(self_send_kind)),
-                        0,
-                        None,
-                        transaction.pools_received(),
-                        memos,
-                    ));
-
-                    // in the case Zennies For Zingo! is active
-                    value_transfers.append(&mut self.create_send_value_transfers(&transaction)?);
-                }
-                TransactionKind::Received => {
-                    // create 1 received value transfer for each pool received to
-                    value_transfers.extend(shielded_pool_value_transfers(
-                        &transaction,
-                        ValueTransferKind::Received,
-                    ));
-                    if !transaction.transparent_coins.is_empty() {
-                        let value: u64 = transaction
-                            .transparent_coins
-                            .iter()
-                            .map(|output| output.value)
-                            .sum();
-                        value_transfers.push(ValueTransfer::from_summary(
-                            &transaction,
-                            ValueTransferKind::Received,
-                            value,
-                            None,
-                            vec![PoolType::TRANSPARENT],
-                            Vec::new(),
-                        ));
-                    }
-                }
-            }
-        }
-
-        Ok(ValueTransfers::new(value_transfers))
-    }
-
     #[must_use]
     pub fn note_summaries<N>(&self, include_spent_notes: bool) -> NoteSummaries
     where
@@ -489,222 +357,21 @@ impl LightWallet {
             })
             .collect()
     }
-
-    /// Provides a list of `ValueTransfers` associated with the sender, or containing the string.
-    pub async fn messages_containing(
-        &self,
-        filter: Option<&str>,
-    ) -> Result<ValueTransfers, SummaryError> {
-        let mut value_transfers = self.value_transfers(true).await?;
-        value_transfers.reverse();
-
-        // Filter out VTs where all memos are empty.
-        value_transfers.retain(|vt| vt.memos.iter().any(|memo| !memo.is_empty()));
-
-        match filter {
-            Some(s) => {
-                value_transfers.retain(|vt| {
-                    if vt.memos.is_empty() {
-                        return false;
-                    }
-
-                    if vt.recipient_address == Some(s.to_string()) {
-                        true
-                    } else {
-                        for memo in &vt.memos {
-                            if memo.contains(s) {
-                                return true;
-                            }
-                        }
-                        false
-                    }
-                });
-            }
-            None => value_transfers.retain(|vt| !vt.memos.is_empty()),
-        }
-
-        Ok(value_transfers)
-    }
-
-    /// TODO: Add Doc Comment Here!
-    pub async fn do_total_memobytes_to_address(
-        &self,
-    ) -> Result<TotalMemoBytesToAddress, SummaryError> {
-        let value_transfers = self.value_transfers(true).await?;
-        let mut memobytes_by_address = HashMap::new();
-        for value_transfer in &value_transfers {
-            if let ValueTransferKind::Sent(SentValueTransfer::Send) = value_transfer.kind {
-                let address = value_transfer
-                    .recipient_address
-                    .clone()
-                    .expect("sent value transfer should always have a recipient_address");
-                let bytes = value_transfer.memos.iter().fold(0, |sum, m| sum + m.len());
-                memobytes_by_address
-                    .entry(address)
-                    .and_modify(|e| *e += bytes)
-                    .or_insert(bytes);
-            }
-        }
-        Ok(TotalMemoBytesToAddress(memobytes_by_address))
-    }
-
-    /// TODO: Add Doc Comment Here!
-    pub async fn do_total_spends_to_address(&self) -> Result<TotalSendsToAddress, SummaryError> {
-        let values_sent_to_addresses = self.value_transfer_by_to_address().await?;
-        let mut by_address_number_sends = HashMap::new();
-        for key in values_sent_to_addresses.0.keys() {
-            let number_sends = values_sent_to_addresses.0[key].len() as u64;
-            by_address_number_sends.insert(key.clone(), number_sends);
-        }
-
-        Ok(TotalSendsToAddress(by_address_number_sends))
-    }
-
-    /// TODO: Add Doc Comment Here!
-    pub async fn do_total_value_to_address(&self) -> Result<TotalValueToAddress, SummaryError> {
-        let values_sent_to_addresses = self.value_transfer_by_to_address().await?;
-        let mut by_address_total = HashMap::new();
-        for key in values_sent_to_addresses.0.keys() {
-            let sum = values_sent_to_addresses.0[key].iter().sum();
-            by_address_total.insert(key.clone(), sum);
-        }
-
-        Ok(TotalValueToAddress(by_address_total))
-    }
-
-    async fn value_transfer_by_to_address(&self) -> Result<ValuesSentToAddress, SummaryError> {
-        let value_transfers = self.value_transfers(false).await?;
-        let mut amount_by_address = HashMap::new();
-        for value_transfer in &value_transfers {
-            if let ValueTransferKind::Sent(SentValueTransfer::Send) = value_transfer.kind {
-                let address = value_transfer
-                    .recipient_address
-                    .clone()
-                    .expect("sent value transfer should always have a recipient_address");
-                amount_by_address
-                    .entry(address)
-                    .and_modify(|e: &mut Vec<u64>| e.push(value_transfer.value))
-                    .or_insert(vec![value_transfer.value]);
-            }
-        }
-
-        Ok(ValuesSentToAddress(amount_by_address))
-    }
-
-    /// Creates value transfers for all notes in a transaction that are sent to another
-    /// recipient.  A value transfer is a group of all notes to a specific receiver in a transaction.
-    /// The value transfer list is sorted by the output index of the notes.
-    fn create_send_value_transfers(
-        &self,
-        transaction: &TransactionSummary,
-    ) -> Result<Vec<ValueTransfer>, KeyError> {
-        let mut value_transfers: Vec<ValueTransfer> = Vec::new();
-        let outgoing_notes = tag_pool(&transaction.outgoing_ironwood_notes, PoolType::IRONWOOD)
-            .chain(tag_pool(
-                &transaction.outgoing_orchard_notes,
-                PoolType::ORCHARD,
-            ))
-            .chain(tag_pool(
-                &transaction.outgoing_sapling_notes,
-                PoolType::SAPLING,
-            ))
-            .collect::<Vec<_>>();
-        let outgoing_coins = &transaction.outgoing_transparent_coins;
-        let mut addresses = HashSet::new();
-
-        outgoing_notes.iter().try_for_each(|&(note, _)| {
-            if note.scope == Scope::External && self.is_wallet_address(&note.recipient)?.is_none() {
-                let encoded_address = note
-                    .recipient_unified_address
-                    .clone()
-                    .unwrap_or(note.recipient.clone());
-                addresses.insert(encoded_address);
-            }
-
-            Ok::<(), KeyError>(())
-        })?;
-        outgoing_coins.iter().try_for_each(|coin| {
-            if self.is_wallet_address(&coin.recipient)?.is_none() {
-                addresses.insert(coin.recipient.clone());
-            }
-
-            Ok::<(), KeyError>(())
-        })?;
-        let mut addresses = addresses.into_iter().collect::<Vec<_>>();
-        addresses.sort();
-        for address in addresses {
-            let outgoing_notes_to_address: Vec<(&OutgoingNoteSummary, PoolType)> = outgoing_notes
-                .iter()
-                .filter(|&&(note, _)| {
-                    let query_address = if let Some(ua) = note.recipient_unified_address.clone() {
-                        ua
-                    } else {
-                        note.recipient.clone()
-                    };
-                    query_address == address
-                })
-                .copied()
-                .collect();
-            let outgoing_coins_to_address: Vec<&OutgoingCoinSummary> = outgoing_coins
-                .iter()
-                .filter(|&coin| coin.recipient.clone() == address)
-                .collect();
-            let value: u64 = outgoing_notes_to_address
-                .iter()
-                .map(|&(note, _)| note.value)
-                .chain(outgoing_coins_to_address.iter().map(|&coin| coin.value))
-                .sum();
-            let memos: Vec<String> = outgoing_notes_to_address
-                .iter()
-                .filter_map(|&(note, _)| note.memo.clone())
-                .collect();
-            let has_notes_in = |pool: PoolType| {
-                outgoing_notes_to_address
-                    .iter()
-                    .any(|&(_, note_pool)| note_pool == pool)
-            };
-            let pools_received = pools_present([
-                !outgoing_coins_to_address.is_empty(),
-                has_notes_in(PoolType::SAPLING),
-                has_notes_in(PoolType::ORCHARD),
-                has_notes_in(PoolType::IRONWOOD),
-            ]);
-            value_transfers.push(ValueTransfer::from_summary(
-                transaction,
-                ValueTransferKind::Sent(SentValueTransfer::Send),
-                value,
-                Some(address),
-                pools_received,
-                memos,
-            ));
-        }
-
-        Ok(value_transfers)
-    }
 }
 
 #[cfg(test)]
 mod tests {
-    use std::str::FromStr as _;
-
-    use pepper_sync::wallet::{
-        IronwoodNote, OrchardNote, OutgoingIronwoodNote, OutputId, WalletTransaction,
-    };
+    use pepper_sync::wallet::{IronwoodNote, OrchardNote, OutputId, WalletTransaction};
     use zcash_primitives::transaction::TxId;
     use zcash_protocol::memo::Memo;
     use zingo_common_components::protocol::ActivationHeights;
     use zingo_status::confirmation_status::ConfirmationStatus;
     use zingo_test_vectors::seeds;
 
-    use crate::ZENNIES_FOR_ZINGO_REGTEST_ADDRESS;
     use crate::config::{ChainType, WalletConfig};
     use crate::mocks::orchard_note::OrchardCryptoNoteBuilder;
     use crate::testutils::default_test_wallet_settings;
     use crate::wallet::LightWallet;
-    use crate::wallet::keys::unified::ReceiverSelection;
-    use crate::wallet::summary::data::{
-        SelfSendValueTransfer, SentValueTransfer, ValueTransferKind,
-    };
 
     /// Message semantics need no network: the ported tests in this module
     /// were libtonode integration tests (the first two ports alone cost
@@ -721,303 +388,6 @@ mod tests {
             },
         )
         .unwrap()
-    }
-
-    fn received(txid_byte: u8, height: u32, memos: &[&str]) -> WalletTransaction {
-        let txid = TxId::from_bytes([txid_byte; 32]);
-        WalletTransaction::new_for_test_with_ironwood_notes(
-            txid,
-            ConfirmationStatus::Confirmed(height.into()),
-            memos
-                .iter()
-                .enumerate()
-                .map(|(index, memo)| {
-                    IronwoodNote::new_for_test(
-                        OutputId::new(txid, index as u32),
-                        zip32::AccountId::ZERO,
-                        zip32::Scope::External,
-                        OrchardCryptoNoteBuilder::default().build(),
-                        Memo::from_str(memo).unwrap(),
-                        None,
-                    )
-                })
-                .collect(),
-            vec![],
-        )
-    }
-
-    fn sent(
-        txid_byte: u8,
-        height: u32,
-        recipient: &zcash_client_backend::address::UnifiedAddress,
-        memo: &str,
-    ) -> WalletTransaction {
-        let txid = TxId::from_bytes([txid_byte; 32]);
-        WalletTransaction::new_for_test_with_ironwood_notes(
-            txid,
-            ConfirmationStatus::Confirmed(height.into()),
-            vec![],
-            vec![OutgoingIronwoodNote::new_for_test(
-                OutputId::new(txid, 0),
-                zip32::AccountId::ZERO,
-                zip32::Scope::External,
-                OrchardCryptoNoteBuilder::default().build(),
-                Memo::from_str(memo).unwrap(),
-                Some(recipient.clone()),
-            )],
-        )
-    }
-
-    /// Migrated from libtonode `fast::filter_empty_messages`.
-    #[tokio::test]
-    async fn filter_empty_messages() {
-        let mut wallet = regtest_wallet(seeds::HOSPITAL_MUSEUM_SEED);
-
-        // Two received notes with empty memos: no messages.
-        wallet
-            .wallet_transactions
-            .insert(TxId::from_bytes([1; 32]), received(1, 10, &["", ""]));
-        assert_eq!(wallet.messages_containing(None).await.unwrap().len(), 0);
-
-        // One real memo alongside an empty one: exactly one message.
-        wallet
-            .wallet_transactions
-            .insert(TxId::from_bytes([2; 32]), received(2, 11, &["Hello", ""]));
-        assert_eq!(wallet.messages_containing(None).await.unwrap().len(), 1);
-    }
-
-    /// Migrated from libtonode `fast::message_thread`.
-    #[tokio::test]
-    async fn message_thread() {
-        // Alice is this wallet; Bob and Charlie are addresses of a foreign
-        // wallet (different seed), exactly as the integration test used the
-        // faucet's addresses.
-        let mut alice_wallet = regtest_wallet(seeds::HOSPITAL_MUSEUM_SEED);
-        let network = ChainType::Regtest(ActivationHeights::default());
-        let alice = alice_wallet
-            .unified_addresses()
-            .values()
-            .next()
-            .unwrap()
-            .encode(&network);
-
-        let mut other_wallet = regtest_wallet(seeds::ABANDON_ART_SEED);
-        let (_, bob) = other_wallet
-            .generate_unified_address(ReceiverSelection::all_shielded(), zip32::AccountId::ZERO)
-            .unwrap();
-        let (_, charlie) = other_wallet
-            .generate_unified_address(ReceiverSelection::all_shielded(), zip32::AccountId::ZERO)
-            .unwrap();
-        let bob_encoded = bob.encode(&network);
-        let charlie_encoded = charlie.encode(&network);
-
-        for transaction in [
-            sent(1, 10, &bob, &format!("Alice->Bob #1\nReply to\n{alice}")),
-            sent(2, 11, &bob, &format!("Alice->Bob #2\nReply to\n{alice}")),
-            received(3, 12, &[&format!("Bob->Alice #2\nReply to\n{bob_encoded}")]),
-            sent(
-                4,
-                13,
-                &charlie,
-                &format!("Alice->Charlie #2\nReply to\n{alice}"),
-            ),
-            received(
-                5,
-                14,
-                &[&format!("Charlie->Alice #2\nReply to\n{charlie_encoded}")],
-            ),
-        ] {
-            alice_wallet
-                .wallet_transactions
-                .insert(transaction.txid(), transaction);
-        }
-
-        let messages_bob = alice_wallet
-            .messages_containing(Some(&bob_encoded))
-            .await
-            .unwrap();
-        let messages_charlie = alice_wallet
-            .messages_containing(Some(&charlie_encoded))
-            .await
-            .unwrap();
-        let all_vts = alice_wallet.value_transfers(true).await.unwrap();
-        let all_messages = alice_wallet.messages_containing(None).await.unwrap();
-
-        assert_eq!(messages_bob.len(), 3);
-        assert_eq!(messages_charlie.len(), 2);
-
-        // ALL MESSAGES (first one should be the oldest one)
-        assert!(
-            all_messages
-                .windows(2)
-                .all(|pair| pair[0].blockheight <= pair[1].blockheight)
-        );
-        // ALL VTS (first one should be the most recent one)
-        assert!(
-            all_vts
-                .windows(2)
-                .all(|pair| pair[0].blockheight >= pair[1].blockheight)
-        );
-    }
-
-    /// Migrated from libtonode `fast::create_send_to_self_with_zfz_active`:
-    /// the assertions are value-transfer KIND classification (a self-send
-    /// yields SendToSelf(Basic). The Zennies-for-Zingo output yields a
-    /// Sent(Send) addressed to the ZFZ address), which is pure summary
-    /// derivation. The proposal/transmission pipeline the integration test
-    /// drove incidentally remains covered by the chain-bound send tests.
-    #[tokio::test]
-    async fn create_send_to_self_with_zfz_active() {
-        let mut wallet = regtest_wallet(seeds::HOSPITAL_MUSEUM_SEED);
-        let network = ChainType::Regtest(ActivationHeights::default());
-
-        let own_orchard_receiver = *wallet
-            .unified_addresses()
-            .values()
-            .next()
-            .unwrap()
-            .orchard()
-            .unwrap();
-        let zcash_client_backend::address::Address::Unified(zfz_unified_address) =
-            zcash_client_backend::address::Address::decode(
-                &network,
-                ZENNIES_FOR_ZINGO_REGTEST_ADDRESS,
-            )
-            .unwrap()
-        else {
-            panic!("ZFZ address must be unified");
-        };
-
-        let txid = TxId::from_bytes([1; 32]);
-        let transaction = WalletTransaction::new_for_test_with_ironwood_notes(
-            txid,
-            ConfirmationStatus::Confirmed(10.into()),
-            vec![],
-            vec![
-                // The send-to-self output: recipient is one of the wallet's
-                // own orchard receivers.
-                OutgoingIronwoodNote::new_for_test(
-                    OutputId::new(txid, 0),
-                    zip32::AccountId::ZERO,
-                    zip32::Scope::External,
-                    OrchardCryptoNoteBuilder::default()
-                        .recipient(own_orchard_receiver)
-                        .build(),
-                    Memo::Empty,
-                    None,
-                ),
-                // The Zennies-for-Zingo output.
-                OutgoingIronwoodNote::new_for_test(
-                    OutputId::new(txid, 1),
-                    zip32::AccountId::ZERO,
-                    zip32::Scope::External,
-                    OrchardCryptoNoteBuilder::default().build(),
-                    Memo::Empty,
-                    Some(zfz_unified_address),
-                ),
-            ],
-        );
-        wallet.wallet_transactions.insert(txid, transaction);
-
-        let value_transfers = wallet.value_transfers(true).await.unwrap();
-
-        assert!(value_transfers.iter().any(|vt| vt.kind
-            == ValueTransferKind::Sent(SentValueTransfer::SendToSelf(
-                SelfSendValueTransfer::Basic
-            ))));
-        assert!(value_transfers.iter().any(|vt| vt.kind
-            == ValueTransferKind::Sent(SentValueTransfer::Send)
-            && vt.recipient_address == Some(ZENNIES_FOR_ZINGO_REGTEST_ADDRESS.to_string())));
-    }
-
-    /// A send value transfer exposes the pools the recipient's outputs
-    /// landed in, letting consumers label which pool a payment was
-    /// delivered to. The funding side (`pools_sent_from`) needs real spend
-    /// links, so it is pinned by the chain-bound tests in libtonode.
-    #[tokio::test]
-    async fn send_exposes_recipient_pools_received() {
-        use zcash_protocol::PoolType;
-
-        let mut alice_wallet = regtest_wallet(seeds::HOSPITAL_MUSEUM_SEED);
-        let mut other_wallet = regtest_wallet(seeds::ABANDON_ART_SEED);
-        let (_, bob) = other_wallet
-            .generate_unified_address(ReceiverSelection::all_shielded(), zip32::AccountId::ZERO)
-            .unwrap();
-
-        alice_wallet
-            .wallet_transactions
-            .insert(TxId::from_bytes([1; 32]), sent(1, 10, &bob, "hi Bob"));
-
-        let value_transfers = alice_wallet.value_transfers(true).await.unwrap();
-
-        let send = value_transfers
-            .iter()
-            .find(|vt| vt.kind == ValueTransferKind::Sent(SentValueTransfer::Send))
-            .unwrap();
-        assert_eq!(send.pools_received, [PoolType::IRONWOOD]);
-        assert!(send.pools_sent_from.is_empty());
-    }
-
-    /// A send-to-self value transfer exposes the pools its value arrived
-    /// into (here: ironwood), letting consumers label pool movements such
-    /// as an orchard -> ironwood migration outside zingolib. The funding
-    /// side (`pools_sent_from`) needs real spend links, so it is pinned by
-    /// the chain-bound shield tests in libtonode instead.
-    #[tokio::test]
-    async fn send_to_self_exposes_pools_received() {
-        use zcash_protocol::PoolType;
-
-        let mut wallet = regtest_wallet(seeds::HOSPITAL_MUSEUM_SEED);
-        let own_orchard_receiver = *wallet
-            .unified_addresses()
-            .values()
-            .next()
-            .unwrap()
-            .orchard()
-            .unwrap();
-
-        let txid = TxId::from_bytes([1; 32]);
-        let transaction = WalletTransaction::new_for_test_with_ironwood_notes(
-            txid,
-            ConfirmationStatus::Confirmed(10.into()),
-            // The self-received output, landing in the ironwood pool.
-            vec![IronwoodNote::new_for_test(
-                OutputId::new(txid, 0),
-                zip32::AccountId::ZERO,
-                zip32::Scope::Internal,
-                OrchardCryptoNoteBuilder::default().build(),
-                Memo::Empty,
-                None,
-            )],
-            // The outgoing view of the same output: recipient is one of
-            // the wallet's own receivers, making the transaction a
-            // send-to-self.
-            vec![OutgoingIronwoodNote::new_for_test(
-                OutputId::new(txid, 0),
-                zip32::AccountId::ZERO,
-                zip32::Scope::External,
-                OrchardCryptoNoteBuilder::default()
-                    .recipient(own_orchard_receiver)
-                    .build(),
-                Memo::Empty,
-                None,
-            )],
-        );
-        wallet.wallet_transactions.insert(txid, transaction);
-
-        let value_transfers = wallet.value_transfers(true).await.unwrap();
-
-        let self_send = value_transfers
-            .iter()
-            .find(|vt| {
-                vt.kind
-                    == ValueTransferKind::Sent(SentValueTransfer::SendToSelf(
-                        SelfSendValueTransfer::Basic,
-                    ))
-            })
-            .unwrap();
-        assert_eq!(self_send.pools_received, [PoolType::IRONWOOD]);
-        assert!(self_send.pools_sent_from.is_empty());
     }
 
     /// Migrated from libtonode
@@ -1392,129 +762,5 @@ mod tests {
             .build();
         let client = LightClient::new_for_test(wallet).await;
         check_client_balances!(client, i: value o: value s: value t: value);
-    }
-
-    /// Migrated from libtonode `slow::by_address_finsight`: the
-    /// memo-bytes-per-address summary accumulates outgoing memo lengths
-    /// keyed by recipient address. Two 1-byte memos then a 4-byte memo.
-    #[tokio::test]
-    async fn by_address_finsight() {
-        use crate::wallet::keys::unified::ReceiverSelection;
-
-        let mut wallet = regtest_wallet(seeds::HOSPITAL_MUSEUM_SEED);
-
-        let mut external_wallet = regtest_wallet(seeds::ABANDON_ART_SEED);
-        let (_, external_ua) = external_wallet
-            .generate_unified_address(ReceiverSelection::all_shielded(), zip32::AccountId::ZERO)
-            .unwrap();
-        let external_ua_encoded = external_ua.encode(&external_wallet.chain_type());
-
-        for (txid_byte, height, memo) in [(1, 4, "1"), (2, 5, "1")] {
-            let transaction = sent(txid_byte, height, &external_ua, memo);
-            wallet
-                .wallet_transactions
-                .insert(transaction.txid(), transaction);
-        }
-        let memobytes = wallet.do_total_memobytes_to_address().await.unwrap();
-        assert_eq!(
-            json::JsonValue::from(memobytes)[&external_ua_encoded].pretty(4),
-            "2".to_string()
-        );
-
-        let transaction = sent(3, 6, &external_ua, "aaaa");
-        wallet
-            .wallet_transactions
-            .insert(transaction.txid(), transaction);
-        let memobytes = wallet.do_total_memobytes_to_address().await.unwrap();
-        assert_eq!(
-            json::JsonValue::from(memobytes)[&external_ua_encoded].pretty(4),
-            "6".to_string()
-        );
-    }
-
-    /// Migrated from libtonode `fast::value_transfers`: a four-output
-    /// memo'd receive aggregates into one value transfer carrying all
-    /// four memos, and the derivation is idempotent and sort-stable
-    /// (the descending ordering is exactly the reverse of ascending).
-    #[tokio::test]
-    async fn value_transfers_aggregation_and_ordering() {
-        use incrementalmerkletree::Position;
-        use orchard::value::NoteValue;
-        use pepper_sync::wallet::OutputId;
-        use std::str::FromStr as _;
-
-        use crate::mocks::orchard_note::OrchardCryptoNoteBuilder;
-
-        let mut wallet = regtest_wallet(seeds::HOSPITAL_MUSEUM_SEED);
-
-        // One received transaction carrying four memo'd notes...
-        let txid = TxId::from_bytes([1; 32]);
-        let notes = (0..4u64)
-            .map(|index| {
-                IronwoodNote::new_for_test(
-                    OutputId::new(txid, u32::try_from(index).unwrap()),
-                    zip32::AccountId::ZERO,
-                    zip32::Scope::External,
-                    OrchardCryptoNoteBuilder::default()
-                        .value(NoteValue::from_raw(5_000))
-                        .build(),
-                    Memo::from_str(&format!("Message #{}", index + 1)).unwrap(),
-                    Some(Position::from(index)),
-                )
-            })
-            .collect::<Vec<_>>();
-        wallet.wallet_transactions.insert(
-            txid,
-            WalletTransaction::new_for_test_with_ironwood_notes(
-                txid,
-                ConfirmationStatus::Confirmed(4.into()),
-                notes,
-                vec![],
-            ),
-        );
-        // ...plus two single-note receives at later heights so the
-        // ordering assertions exercise a non-trivial sort.
-        for (txid_byte, height) in [(2u8, 5u32), (3, 6)] {
-            let txid = TxId::from_bytes([txid_byte; 32]);
-            let note = IronwoodNote::new_for_test(
-                OutputId::new(txid, 0),
-                zip32::AccountId::ZERO,
-                zip32::Scope::External,
-                OrchardCryptoNoteBuilder::default()
-                    .value(NoteValue::from_raw(1_000 * u64::from(txid_byte)))
-                    .build(),
-                Memo::Empty,
-                Some(Position::from(u64::from(txid_byte))),
-            );
-            wallet.wallet_transactions.insert(
-                txid,
-                WalletTransaction::new_for_test_with_ironwood_notes(
-                    txid,
-                    ConfirmationStatus::Confirmed(height.into()),
-                    vec![note],
-                    vec![],
-                ),
-            );
-        }
-
-        let value_transfers = wallet.value_transfers(true).await.unwrap();
-        let value_transfers1 = wallet.value_transfers(true).await.unwrap();
-        let value_transfers2 = wallet.value_transfers(true).await.unwrap();
-        let mut value_transfers3 = wallet.value_transfers(false).await.unwrap();
-        let mut value_transfers4 = wallet.value_transfers(false).await.unwrap();
-
-        let four_memo_transfer = value_transfers
-            .iter()
-            .find(|transfer| transfer.txid == TxId::from_bytes([1; 32]))
-            .unwrap();
-        assert_eq!(four_memo_transfer.memos.len(), 4);
-
-        value_transfers3.reverse();
-        value_transfers4.reverse();
-
-        assert_eq!(value_transfers, value_transfers1);
-        assert_eq!(value_transfers, value_transfers2);
-        assert_eq!(value_transfers, value_transfers3);
-        assert_eq!(value_transfers, value_transfers4);
     }
 }
